+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - CommonJS support;
 - `.abs` property to get absolute value of the rational number;
+- `.round()`, `.ceil()`, and `floor()` method to perform round operations on rational numbers;
 
 ## 0.1.0 (2021-07-29)